backend/requirements.txt
alembic==1.12.0
attrs==23.1.0
bidict==0.22.1
blinker==1.6.2
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.0
click==8.1.7
colorama==0.4.6
cors==1.0.1
coverage==6.0
cryptography==41.0.4
filelock==3.12.4
Flask==3.0.0
Flask-Cors==4.0.0
Flask-JWT-Extended==4.5.3
Flask-Migrate==4.0.5
Flask-Script==2.0.6
Flask-SQLAlchemy==3.1.1
future==0.18.3
gevent==23.9.1
greenlet==3.0.0
h11==0.14.0
idna==3.4
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
jwt==1.3.1
Mako==1.2.4
MarkupSafe==2.1.3
mysqlclient==2.2.0
packaging==23.2
pluggy==1.3.0
py==1.11.0
pycparser==2.21
PyJWT==2.8.0
PySocks==1.7.1
pytest==6.2.5
python-dotenv==1.0.0
python-engineio==4.7.1
python-socketio==5.9.0
requests==2.31.0
requests-file==1.5.1
simple-websocket==1.0.0
six==1.16.0
SQLAlchemy==2.0.21
tldextract==3.6.0
toml==0.10.2
typing_extensions==4.8.0
urllib3==2.0.6
Werkzeug==3.0.0
wsproto==1.2.0
zope.event==5.0
zope.interface==6.1
